antenna model v1
The matlab script provides the array pattern for an array of patch antenna elements arranged in a rectangular fashion along the x and y dimensions (linear arrays are special cases). The array pattern is saved in "arraypattern.mat". The pattern of the patch element is available in the data file "realized_gain.csv".